+++ extensions/calDotCom/index.ts
@@ -1,6 +1,11 @@
 import { type Extension } from '@awell-health/extensions-core'
 import { AuthorType, Category } from '@awell-health/extensions-core'
-import { bookAppointment, getBooking } from './actions'
+import {
+  bookAppointment,
+  getBooking,
+  updateBooking,
+  deleteBooking,
+} from './actions'
 import { settings } from './settings'
 import { webhooks } from './webhooks'
 
@@ -16,6 +21,8 @@ export const CalDotCom: Extension = {
   actions: {
     bookAppointment,
     getBooking,
+    updateBooking,
+    deleteBooking,
   },
   webhooks,
   settings,