+++ extensions/calDotCom/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Cal.com changelog
 
+## September 4, 2023
+
+- add updateBooking and deleteBooking actions
+- update and refactor structure of getBooking and bookAppointment actions
+
 ## July 4, 2023
 
 - Get booking action: add validation that bookingId and apiKey are not empty